fix: Não é possível registrar propostas de nomes iguais, tipo de atributo de tempo definido para long

+++ src/main/java/com/guardioes/propostas/web/dto/VotacaoInitDto.java
@@ -12,5 +12,5 @@ import lombok.Setter;
 public class VotacaoInitDto {
     private String propostaTitulo;
     private String funcionarioCpf;
-    private int tempo=1 ;
+    private long tempo = 1 ;
 }
